[warnings] UrlImplWinInet.cpp: fix narrowing conversion warnings

+++ src/cinder/UrlImplWinInet.cpp
@@ -231,26 +231,26 @@ void IStreamUrlImplWinInet::fillBuffer( int wantBytes ) const
 
 void IStreamUrlImplWinInet::IORead( void *dest, size_t size )
 {
-	fillBuffer( size );
-	
+	fillBuffer( static_cast<int>( size ) );
+
 	// check if theres data in the buffer - if not fillBuffer() either errored or EOF
 	if( bufferDataRemaining() < (off_t)size )
 		throw StreamExc();
 
 	memcpy( dest, mBuffer.data() + mBufferOffset, size );
-	mBufferOffset += size;
+	mBufferOffset += static_cast<int>( size );
 }
 
 size_t IStreamUrlImplWinInet::readDataAvailable( void *dest, size_t maxSize )
 {
-	fillBuffer( maxSize );
-	
+	fillBuffer( static_cast<int>( maxSize ) );
+
 	if( bufferDataRemaining() < (off_t)maxSize )
 		maxSize = bufferDataRemaining();
-		
+
 	memcpy( dest, mBuffer.data() + mBufferOffset, maxSize );
-	
-	mBufferOffset += maxSize;
+
+	mBufferOffset += static_cast<int>( maxSize );
 	return maxSize;
 }
